+++ src/ArrayMicRefreshment.Asr/AsrEngineKind.cs
@@ -2,6 +2,7 @@ namespace ArrayMicRefreshment.Asr;
 
 public enum AsrEngineKind
 {
+    Qwen3Asr,
     SenseVoice,
     FireRedCtc,
 }